+++ app/static/js/modules/utils/popup.js
