+++ app/templates/adapters.html
@@ -54,9 +54,9 @@
     <div id="alert-message" class="popup-body"></div>
 </div>
 
-<script type="module" src="/static/js/modules/popup.js"></script>
-<script type="module" src="/static/js/modules/a_controlls.js"></script>
-<script type="module" src="/static/js/modules/a_form.js"></script>
+<script type="module" src="/static/js/modules/utils/popup.js"></script>
+<script type="module" src="/static/js/modules/adapter/control.js"></script>
+<script type="module" src="/static/js/modules/adapter/form.js"></script>
 
 <script type="module" src="/static/js/adapters.js"></script>
 {% endblock %}
